+++ src/components/webinaire/form/pages/index.ts
@@ -1,0 +1,6 @@
+import { OtherDataForm } from './OtherDataForm';
+import { UserInfoForm } from './UserInfoForm';
+
+const Pages = [UserInfoForm, OtherDataForm];
+
+export { Pages };
